+++ src/ops_data_store/cli/config/__init__.py
@@ -11,6 +11,19 @@ config = Config()
 logger = logging.getLogger("app")
 
 
+@app.command(help="Check current app configuration is valid.")
+def check() -> None:
+    """Validate current app configuration."""
+    try:
+        logger.info("Checking app config.")
+        config.validate()
+    except RuntimeError as e:
+        logger.error(e, exc_info=True)
+        typer.echo(e)
+        raise typer.Abort() from e
+
+    logger.info("App config ok.")
+    print("Ok. Configuration valid.")
 
 
 @app.command(help="Display the current app configuration.")