+++ tests/ops_data_store_tests/cli/test_cli_config.py
@@ -9,6 +9,34 @@ from ops_data_store.cli import app as cli
 from ops_data_store.config import Config
 
 
+class TestCliConfigCheck:
+    """Tests for `config check` CLI command."""
+
+    def test_ok(self, caplog: pytest.LogCaptureFixture, fx_cli_runner: CliRunner) -> None:
+        """Returns ok message."""
+        result = fx_cli_runner.invoke(app=cli, args=["config", "check"])
+
+        assert "Checking app config." in caplog.text
+        assert "App config ok." in caplog.text
+
+        assert result.exit_code == 0
+        assert "Ok. Configuration valid." in result.output
+
+    def test_invalid(self, caplog: pytest.LogCaptureFixture, fx_cli_runner: CliRunner, mocker: MockerFixture) -> None:
+        """Returns error message."""
+        mocker.patch.object(
+            target=Config, attribute="validate", side_effect=RuntimeError("Required config option `DB_DSN` not set.")
+        )
+
+        result = fx_cli_runner.invoke(app=cli, args=["config", "check"])
+
+        assert "Checking app config." in caplog.text
+        assert "Required config option `DB_DSN` not set." in caplog.text
+
+        assert result.exit_code == 1
+        assert "Required config option `DB_DSN` not set." in result.output
+
+
 class TestCliConfigShow:
     """Tests for `config show` CLI command."""